Moves algeron facets to .api.algeron
