+++ src/main/java/org/jboss/forge/arquillian/api/algeron/AlgeronConsumer.java
@@ -1,4 +1,4 @@
-package org.jboss.forge.arquillian.testframework.algeron;
+package org.jboss.forge.arquillian.api.algeron;
 
 import org.jboss.forge.addon.dependencies.builder.DependencyBuilder;
 import org.jboss.forge.arquillian.api.algeron.AlgeronSetupFacet;
