+++ src/main/java/org/jboss/forge/arquillian/command/algeron/provider/AlgeronSetupProvider.java
@@ -15,7 +15,7 @@ import org.jboss.forge.addon.ui.util.Metadata;
 import org.jboss.forge.arquillian.api.TestFrameworkFacet;
 import org.jboss.forge.arquillian.command.algeron.AbstractAlgeronCommand;
 import org.jboss.forge.arquillian.container.model.ContractProviderLibrary;
-import org.jboss.forge.arquillian.testframework.algeron.AlgeronProvider;
+import org.jboss.forge.arquillian.api.algeron.AlgeronProvider;
 
 import javax.inject.Inject;
 import java.util.Arrays;
