+++ src/main/java/org/jboss/forge/arquillian/command/algeron/publisher/AlgeronPublisher.java
@@ -2,13 +2,13 @@ package org.jboss.forge.arquillian.command.algeron.publisher;
 
 import org.jboss.forge.addon.ui.command.UICommand;
 
-enum AlgeronPublishers {
+enum AlgeronPublisher {
 
     FOLDER(AddAlgeronFolderPublisher.class), URL(AddAlgeronUrlPublisher.class), GIT(AddAlgeronGitPublisher.class);
 
     private Class<? extends UICommand> implementingCommand;
 
-    AlgeronPublishers(Class<? extends UICommand> implementingCommand) {
+    AlgeronPublisher(Class<? extends UICommand> implementingCommand) {
         this.implementingCommand = implementingCommand;
     }
